chore: Kpi update with loading  and error handling , handover update end cleanup
