+++ client/packages/portal-pages/src/pages/context-page/components/kpis/LowCarbon.tsx
@@ -1,6 +1,14 @@
-import { KpiCard } from "@equinor/portal-ui"
-
+import { KpiCard } from '@equinor/portal-ui';
 
 export const LowCarbon = () => {
-    return <KpiCard title="Low Carbon" items={[{title: "CO2 intensity", value: "14.48"},{title: "CO2 emission", value: "8.30"}]}/>
-}
+	return (
+		<KpiCard
+			isLoading={false}
+			title="Low Carbon"
+			items={[
+				{ title: 'CO2 intensity', value: '14.48' },
+				{ title: 'CO2 emission', value: '8.30' },
+			]}
+		/>
+	);
+};
