+++ client/packages/portal-ui/src/lib/kpis/KpiItem.tsx
@@ -1,39 +1,34 @@
-
-
-
 import styled from 'styled-components';
 import { tokens } from '@equinor/eds-tokens';
 import { KpiItem } from './types';
 
 const StyledValue = styled.div`
-    color: ${tokens.colors.text.static_icons__default.hex};
-    font-size: 20px;
-    line-height: 24px;
+	color: ${tokens.colors.text.static_icons__default.hex};
+	font-size: 20px;
+	line-height: 24px;
 `;
 
 const StyledValueWrapper = styled.div`
-    display: flex;
-    flex-direction: row;
-    align-items: center;
-    gap: 0.2em;
+	display: flex;
+	flex-direction: row;
+	align-items: center;
+	gap: 0.2em;
 `;
 
 const StyledTitle = styled.div`
-    color: ${tokens.colors.text.static_icons__tertiary.hex};
-    font-size: 12px;
-    line-height: 16px;
+	color: ${tokens.colors.text.static_icons__tertiary.hex};
+	font-size: 12px;
+	line-height: 16px;
 `;
 
 const StyledStatusCardItem = styled.div`
-    min-height: 40;
-    min-width: 70;
-    width: fit-content;
+	min-height: 40px;
+	width: fit-content;
 `;
 
-
-interface StatusBarItemProps {
+type StatusBarItemProps = {
 	item: KpiItem;
-}
+};
 export function KpiCardItem({ item }: StatusBarItemProps) {
 	return (
 		<StyledStatusCardItem title={item.description} key={item.title}>
@@ -44,4 +39,3 @@ export function KpiCardItem({ item }: StatusBarItemProps) {
 		</StyledStatusCardItem>
 	);
 }
-
