+++ client/packages/portal-ui/src/lib/progress-loader/ProgressLoader.tsx
@@ -1,4 +1,4 @@
-import { StarProgress, Typography } from '@equinor/eds-core-react';
+import { CircularProgress, Typography } from '@equinor/eds-core-react';
 import styled from 'styled-components';
 
 const StyledWrapper = styled.div`
@@ -9,19 +9,15 @@ const StyledWrapper = styled.div`
 	align-items: center;
 `;
 
-const StyledProgress = styled(StarProgress)`
-	height: auto;
-	width: 85px;
-`;
-
 type ProgressLoaderProps = {
 	title: string;
+	size?: 16 | 24 | 32 | 40 | 48;
 };
 
-export function PortalProgressLoader({ title }: ProgressLoaderProps) {
+export function ProgressLoader({ title, size }: ProgressLoaderProps) {
 	return (
 		<StyledWrapper>
-			<StyledProgress />
+			<CircularProgress size={size} />
 			<Typography variant="h5"> {title}</Typography>
 		</StyledWrapper>
 	);
