+++ client/packages/portal-ui/src/lib/progress-loader/index.ts
@@ -1,1 +1,2 @@
-export * from "./PortalProgressLoader"
+export * from './PortalProgressLoader';
+export * from './ProgressLoader';
